Conflicts:
	src/ServiceManagement/RecoveryServices/Commands.RecoveryServices.Test/ScenarioTests/RecoveryServicesTests.ps1
	src/ServiceManagement/RecoveryServices/Commands.RecoveryServices/Service/CreateAzureSiteRecoveryProtectionProfileObject.cs
	src/ServiceManagement/RecoveryServices/Commands.RecoveryServices/Service/SetAzureSiteRecoveryProtectionEntity.cs

+++ src/ServiceManagement/RecoveryServices/Commands.RecoveryServices/Service/GetAzureSiteRecoveryRecoveryPlanFile.cs
@@ -14,9 +14,9 @@
 
 using System;
 using System.Collections.Generic;
+using System.IO;
 using System.Management.Automation;
 using Microsoft.Azure.Commands.RecoveryServices.SiteRecovery;
-using Microsoft.WindowsAzure;
 using Microsoft.WindowsAzure.Management.SiteRecovery.Models;
 
 namespace Microsoft.Azure.Commands.RecoveryServices
@@ -80,7 +80,12 @@ namespace Microsoft.Azure.Commands.RecoveryServices
         /// </summary>
         private void GetRecoveryPlanFile()
         {
-            RecoveryPlanXmlOuput recoveryPlanXmlOuput = 
+            if (Directory.Exists(this.Path))
+            {
+                throw new ArgumentException("The input path is a directory. Please provide file path. Check the examples of the commandlet", "Path");
+            }
+
+            RecoveryPlanXmlOuput recoveryPlanXmlOuput =
                 RecoveryServicesClient.GetAzureSiteRecoveryRecoveryPlanFile(this.Id);
             System.IO.File.WriteAllText(this.Path, recoveryPlanXmlOuput.RecoveryPlanXml);
         }
